add CLI for saving image preview of layout (#1)
Adds an `ergodox-screenshot-layout` CLI for saving an image preview of a keyboard layout.

The tool uses [puppeteer][] to visit the [Ergodox Configurator tool][configurator], import your keyboard layout specified by path, and save a screenshot.

Example usage: `ergodox-screenshot-layout -f MDErgo1-Default.json` will save a preview of the layout to `screenshot.png`. The path to the keyboard layout file is specified with the `-f` option. The path to save the screenshot to can be specified with the `-d` option. The CLI also supports the `--headless` flag which can be set to false to configure puppeteer to not run chrome in headless mode (for debugging).

[puppeteer]: https://github.com/GoogleChrome/puppeteer
[configurator]: https://input.club/configurator-ergodox/